trivial: move clone firmware shell code into common script

+++ contrib/ci/get_test_firmware.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 
-#clone fwupd-test-firmware
-rm -rf fwupd-test-firmware
-git clone https://github.com/fwupd/fwupd-test-firmware
-#set up build tests to work
-cp fwupd-test-firmware/ci-tests/* . -R
-export G_TEST_SRCDIR=`pwd`/fwupd-test-firmware/installed-tests
+if [ "$CI_NETWORK" = "true" ]; then
+        #clone fwupd-test-firmware
+        rm -rf fwupd-test-firmware
+        git clone https://github.com/fwupd/fwupd-test-firmware
+        #set up build tests to work
+        cp fwupd-test-firmware/ci-tests/* . -R
+        root=$(pwd)
+        export G_TEST_SRCDIR=${root}/fwupd-test-firmware/installed-tests
+fi
